pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "folio"
version = "0.1.0"
description = "Your consulting portfolio, searchable and AI-ready."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Johnny Oh"}
]

dependencies = [
    "click>=8.0",
    "markitdown>=0.1",
    "pdf2image>=1.16",
    "Pillow>=9.0",
    "anthropic>=0.40",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
folio = "folio.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
